refactor(core): split geometry paths, axis training, and panel chrome
Extract coord-flip and mark-count helpers from geometry.ts; line vs area
path builders from geometry-paths; axis types/domain from scale-axis-train;
and labs/display/legends from panel-layout-chrome. Facades keep import
paths stable. Extend scale-training characterization for continuousDomainOf.
