+++ packages/core/src/pipeline/geometry-paths-line.ts
@@ -1,0 +1,65 @@
+/**
+ * Line path geometry batch builder.
+ */
+import type { PathsBatch } from "../scene.js";
+
+import type { LayerFrame, PipelineWarning, ResolvedColorScale } from "./types.js";
+import { colorOf } from "./types.js";
+import type { Frame } from "./geometry-shared.js";
+import { DEFAULT_LINEWIDTH, bucketByGroup, positionOf, xSortKey } from "./geometry-shared.js";
+
+export function lineBatch(
+  frame: LayerFrame,
+  fx: Frame,
+  color: ResolvedColorScale | null,
+  warnings: PipelineWarning[],
+): PathsBatch | null {
+  const { binding } = frame;
+  const subpaths = bucketByGroup(frame, fx, null, warnings);
+  if (subpaths.length === 0) return null;
+  const sortKey = xSortKey(frame, fx);
+  for (const rows of subpaths) rows.sort((a, b) => sortKey(a) - sortKey(b));
+
+  let total = 0;
+  for (const rows of subpaths) total += rows.length;
+  const positions = new Float32Array(total * 2);
+  const rowIndex = new Uint32Array(total);
+  const pathOffsets = new Uint32Array(subpaths.length + 1);
+  const strokes: (string | null)[] = [];
+  let cursor = 0;
+  for (let s = 0; s < subpaths.length; s++) {
+    pathOffsets[s] = cursor;
+    const rows = subpaths[s]!;
+    for (const row of rows) {
+      const tx = positionOf(fx.xScale, frame.xNumeric, frame.xValues, row);
+      const ty = positionOf(fx.yScale, frame.yNumeric, null, row);
+      positions[cursor * 2] = tx * fx.innerWidth;
+      positions[cursor * 2 + 1] = fx.innerHeight - ty * fx.innerHeight;
+      rowIndex[cursor] = frame.rowIndex[row]!;
+      cursor++;
+    }
+    let stroke: string | null = binding.color.constant;
+    if (color !== null && (frame.colorValues !== null || binding.color.scaledConstant !== null)) {
+      const first = rows[0]!;
+      const value =
+        frame.colorValues === null ? binding.color.scaledConstant! : frame.colorValues[first]!;
+      stroke = colorOf(color, value);
+    }
+    strokes.push(stroke);
+  }
+  pathOffsets[subpaths.length] = cursor;
+
+  const params = binding.layer.geom === "line" ? (binding.layer.params ?? {}) : {};
+  return {
+    kind: "paths",
+    layerIndex: binding.index,
+    panelIndex: 0,
+    positions,
+    rowIndex,
+    pathOffsets,
+    strokes,
+    linewidth: params.linewidth ?? DEFAULT_LINEWIDTH,
+    alpha: params.alpha ?? 1,
+    curve: params.curve ?? "linear",
+  };
+}
